backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==2.0.2
shap==0.44.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic==2.5.0